requirements.txt
python-telegram-bot==13.15
google-api-python-client==2.56.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
